requirements.txt
Arpeggio==2.0.2
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dogpile.cache==1.4.0
grpcio==1.66.2
idna==3.10
iso8601==2.1.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keystoneauth1==5.11.1
openstacksdk==4.6.0
os-service-types==1.7.0
parver==0.5
pbr==6.1.1
platformdirs==4.3.8
protobuf==4.25.8
psutil==7.0.0
pulumi==3.181.0
pulumi_openstack==5.1.0
pycparser==2.22
PyYAML==6.0.2
requests==2.32.4
requestsexceptions==1.4.0
semver==3.0.4
stevedore==5.4.1
typing_extensions==4.14.1
urllib3==2.5.0
